The end of an era: hide configuration of the superuser name as an Easter egg.
It was an interesting experiment, often fun, often annoying, often tiring.
We stretched the boundaries of what's hardcoded and what's not in a Linux
system, but as of 2014 it just takes too much effort to swim against the
stream on that, for little to no practical gain.

Keeping this as a hidden option in the 015 release CD because of all the 
effort Lucas went through to make it work in this release. From here on
we should start to simplify recipes and scripts by assuming that user 0
is always called "root".